1
1

+++ Assets/Scripts/EnemyData.cs
@@ -1,6 +1,6 @@
 class EnemyData
 {
     public float SpawnTime;
-    public float EnemyKind;
+    public int EnemyKind;
     public int Spawners;
 }
